feat: server-side generate Learn
